+++ src/calculation-model/documentation.tsx
@@ -1,10 +1,14 @@
 import React, { Component } from 'react'
+/*
 import ReactMarkdown from 'react-markdown'
 import remarkMath from 'remark-math'
 import remarkGfm from 'remark-gfm'
 import rehypeKatex from 'rehype-katex'
-import 'katex/dist/katex.min.css'
+import 'katex/dist/katex.min.css'7
 import documentationMd from './documentation.md'
+*/
+import { secureLink } from '../helpers';
+import { strings } from '../constants/textData'
 
 interface ICalcModDocProps {
 
@@ -23,12 +27,15 @@ export class CalculationModelDocumentation extends Component<ICalcModDocProps, I
   }
 
   componentDidMount() {
-    fetch(documentationMd).then(res => res.text()).then(text => this.setState({ md: text }));
+    //fetch(documentationMd).then(res => res.text()).then(text => this.setState({ md: text }));
   }
 
   render() {
-    const { md } = this.state;
+    // const { md } = this.state;
     return (
+      // Currently, documentation resides in OverLeaf which is not sustainable. It should ideally be moved to this git repo.
+      <div>Detailed documentation for the calculation model is available on {secureLink(strings.pdfDocsLink, "GitHub")}.</div>
+      /*
       <div id="calculation-model-documentation">
           <ReactMarkdown 
             children={md}
@@ -36,6 +43,7 @@ export class CalculationModelDocumentation extends Component<ICalcModDocProps, I
             rehypePlugins={[rehypeKatex]}
           />
       </div>
+      */
     ) 
   }
 }